refactor: Replace int instanceId parameters with InventoryItemInstanceId type
Replace all occurrences of `int instanceId` method parameters with the strongly-typed `InventoryItemInstanceId id` throughout the inventory system. This improves type safety and makes the API more explicit.

Changes:
- Update method signatures across Inventory, DraggingItem, and UI components
- Change internal NativeHashMap from <int, int> to <ulong, int> for compatibility
- Update all Dictionary and collection types to use InventoryItemInstanceId
- Fix all tests to use explicit InventoryItemInstanceId casts
- Update Items.NextItemInstanceId to return InventoryItemInstanceId type

+++ Assets/Inventory/TestInventoryView.cs
@@ -62,7 +62,7 @@ public class TestInventoryView : MonoBehaviour
             // Build model item definition from UI shape
             var instanceId = _items.NextItemInstanceId;
             var itemDef = new ItemDefinition(guid, ui.Shape.ToImmutableGridShape());
-            var rotation = rotations[instanceId % rotations.Length];
+            var rotation = rotations[instanceId % (ulong)rotations.Length];
             var candidate = new InventoryItem(instanceId, itemDef, rotation, new int2(-1, -1));
 
             if (!_inventory.TryAutoPlaceItem(candidate, out _))
